requirements.txt
asgiref
Django
django-cors-headers
django-environ
django-filter
djangorestframework
djangorestframework-simplejwt
pandas
numpy
psycopg2-binary
PyJWT
python-dateutil
pytz
scapy
six
sqlparse
urllib3
whitenoise
setuptools
wheel
